requirements.txt
numpy
pillow
opencv-python
tqdm
scikit-learn
timm
torch
torchvision
